Split IntakeSource modules
