+++ lumen/sources/intake.py
@@ -84,67 +84,3 @@ class IntakeSource(IntakeBaseSource):
             entries = {entry.name: entry for entry in cfg.pop('data_sources')}
             self.cat = intake.catalog.Catalog(**cfg)
             self.cat._entries = entries
-
-
-class IntakeBaseSQLSource(IntakeBaseSource):
-
-    # Declare this source supports SQL transforms
-    _sql = True
-
-    @cached()
-    def get(self, table, **query):
-        '''
-        Applies SQL Transforms, creating new temp catalog on the fly
-        and querying the database.
-        '''
-        dask = query.pop('__dask', self.dask)
-        sql_transforms = query.pop('sql_transforms', [])
-
-        try:
-            source = self.cat[table]
-        except KeyError:
-            raise KeyError(
-                f"'{table}' table could not be found in Intake catalog. "
-                f"Available tables include: {list(self.cat)}."
-            )
-        sql_expr = source._sql_expr
-        for sql_transform in sql_transforms:
-            sql_expr = sql_transform.apply(sql_expr)
-        new_source = type(source)(**dict(source._init_args, sql_expr=sql_expr))
-        df = self._read(new_source)
-        return df if dask or not hasattr(df, 'compute') else df.compute()
-
-
-class IntakeDremioSource(IntakeBaseSQLSource):
-
-    cert = param.String(default="Path to certificate file")
-
-    dask = param.Boolean(default=False, doc="""
-        Whether to return a dask DataFrame.""")
-
-    uri = param.String(doc="URI of the catalog file.")
-
-    tls = param.Boolean(default=False, doc="Enable encryption")
-
-    username = param.String(default=None, doc="Dremio username")
-
-    password = param.String(default=None, doc="Dremio password or token")
-
-    source_type = 'intake_dremio'
-
-    def __init__(self, **params):
-        from intake_dremio.dremio_cat import DremioCatalog
-        super().__init__(**params)
-        self.cat = DremioCatalog(
-            self.uri, cert=self.cert, tls=self.tls, username=self.username,
-            password=self.password
-        )
-
-
-class IntakeSQLSource(IntakeBaseSQLSource, IntakeSource):
-    """
-    Intake source specifically for SQL sources.
-    Allows for sql transformations to be applied prior to querying the source.
-    """
-
-    source_type = 'intake_sql'
